Separate JNI internal functionalities to internal.h
